+++ rosflight_utils/include/rosflight_utils/turbomath.h
