+++ rosflight_utils/src/simple_pid.cpp
@@ -1,6 +1,6 @@
-#include "rosflight_common/simple_pid.h"
+#include <rosflight_utils/simple_pid.h>
 
-namespace rosflight_common
+namespace rosflight_utils
 {
 //
 // Basic initialization
